mi_api/requirements.txt
flask
flask-sqlalchemy
psycopg2-binary
flask-jwt-extended
python-dotenv
flasgger
flask-cors
werkzeug
